LLM/requirements.txt
requests~=2.32.5
pandas~=2.3.3
fastapi~=0.119.0
pydantic~=2.12.3
uvicorn~=0.38.0
langchain~=1.0.0
langchain_chroma~=1.0.0
langchain-community~=0.4
langchain-huggingface~=1.0.0
langchain-text-splitters~=1.0.0
torch~=2.9.0
torchvision~=0.24.0
torchaudio~=2.9.0
sentence-transformers~=5.1.1
chromadb~=1.2.0
ollama~=0.6.0
google-adk~=1.16.0
litellm~=1.78.4
pypdf~=6.1.1
